Add files via upload
This is an improved version of my personal website. In this project, I reviewed my understanding of CSS3. I appreciate the efforts I expended to relearn. In conclusion, I used HTML5 and CSS3 to complete this amazing project, and I will work hard to understand more and build better projects with awesome UI/UX.😎😎😎